requirements.txt
celery==5.2.3
Django==3.2.12
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.3
psutil==5.9.0
redis==4.4.0
django-redis==5.2.0
user-agents==2.2.0
daphne==4.0.0
channels-redis==4.0.0
django-restql==0.15.2
requests==2.28.1
xlwt==1.3.0
openpyxl==3.1.0
cron-descriptor==1.2.35
